# docs/reference/StockWatcher-Web-Internal-Test-Handoff-20260807/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StockWatcher Web Internal Test API
  version: 1.0.0-draft
  description: Contract for the headless StockWatcher worker and authenticated internal Web UI. Business time is
    Asia/Shanghai. This API never exposes credentials.
servers:
- url: https://{domain}
  variables:
    domain:
      default: stock.example.com
...
paths:
...
  /api/v1/history:
    get:
      tags:
      - history
      security:
      - SessionCookie: []
      summary: Candidate snapshot history
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: from
        in: query
        schema:
          type: string
          format: date
      - name: to
        in: query
        schema:
          type: string
          format: date
      - name: code
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/alerts:
    get:
      tags:
      - history
      security:
      - SessionCookie: []
      summary: Alert history
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: from
        in: query
        schema:
          type: string
          format: date
      - name: to
        in: query
        schema:
          type: string
          format: date
      - name: trigger_type
        in: query
        schema:
          type: string
          enum:
          - scheduled-09:45
          - scheduled-14:45
          - intraday
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAlerts'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/outcomes:
    get:
      tags:
      - outcomes
      security:
      - SessionCookie: []
      summary: Read-only next-trading-day candidate outcome review
      parameters:
      - name: range
        in: query
        schema:
          type: string
          enum:
          - week
          - month
          - all
          default: month
      responses:
        '200':
          description: Sanitized outcome statistics and records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutcomeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sw_session
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type:
        - string
        - 'null'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Role or CSRF check failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    OutcomeResponse:
      type: object
      required:
      - range
      - summary
      - morning
      - afternoon
      - portfolio
      - records
      - backfill
      properties:
        range:
          type: string
          enum:
          - week
          - month
          - all
        summary:
          type: object
          additionalProperties: true
        morning:
          type: object
          additionalProperties: true
        afternoon:
          type: object
          additionalProperties: true
        portfolio:
          type: object
          additionalProperties: true
        records:
          type: array
          items:
            type: object
            additionalProperties: true
        backfill:
          type: object
          additionalProperties: true
      additionalProperties: false
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type:
              - object
              - 'null'
              additionalProperties: true
      additionalProperties: false
...
    Candidate:
      type: object
      required:
      - rank
      - code
      - name
      - level
      - is_formal
      - is_supplement
      - price
      - change_pct
      - sector_name
      - sector_type
      - total_score
      - source_ts
      properties:
        rank:
          type: integer
          minimum: 1
          maximum: 3
        code:
          type: string
        name:
          type: string
        level:
          type: string
          enum:
          - 强
          - 中
          - 近
        is_formal:
          type: boolean
        is_supplement:
          type: boolean
        price:
          type: number
        change_pct:
          type: number
        velocity_pct:
          type: number
        velocity_available:
          type: boolean
        velocity_1m_pct:
          type:
          - number
          - 'null'
        velocity_3m_pct:
          type:
          - number
          - 'null'
        velocity_5m_pct:
          type:
          - number
          - 'null'
        sector_code:
          type: string
        sector_name:
          type: string
        sector_type:
          type: string
          enum:
          - industry
          - concept
        price_score:
          type: number
        sector_score:
          type: number
        trend_score:
          type: number
        fund_score:
          type: number
        core_score:
          type: number
        total_score:
          type: number
        penalty:
          type: number
        fund_label:
          type: string
        trend_label:
          type: string
        data_completeness:
          type: number
          minimum: 0
          maximum: 1
        sector_gate_passed:
          type: boolean
        reasons:
          type: array
          items:
            type: string
        source_ts:
          type: string
          format: date-time
      additionalProperties: false
...
    HistoryItem:
      type: object
      required:
      - snapshot_id
      - generated_at
      - health
      - overall_weak
      - candidates
      properties:
        snapshot_id:
          type: integer
        generated_at:
          type: string
          format: date-time
        health:
          type: string
        overall_weak:
          type: boolean
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
      additionalProperties: false
    Alert:
      type: object
      required:
      - alert_id
      - snapshot_id
      - displayed_at
      - trigger_type
      - decision
      - channel
      properties:
        alert_id:
          type: integer
        snapshot_id:
          type: integer
        displayed_at:
          type: string
          format: date-time
        trigger_type:
          type: string
          enum:
          - scheduled-09:45
          - scheduled-14:45
          - intraday
        decision:
          type: string
        channel:
          type: string
        triggering_codes:
          type: array
          items:
            type: string
        strength:
          type:
          - number
          - 'null'
        funds_unconfirmed:
          type:
          - boolean
          - 'null'
      additionalProperties: false
...
    PagedHistory:
      type: object
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistoryItem'
        next_cursor:
          type:
          - string
          - 'null'
      additionalProperties: false
    PagedAlerts:
      type: object
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        next_cursor:
          type:
          - string
          - 'null'
      additionalProperties: false
...